Updates
Removed jQuery fadein, doesnt work on firefox replaced with a webkit transition.  Added a class to the li with the bbb ranking so that they would all be the same height.  changed the nav links to display: none until we have functioning links.  changed the footer links to display: none until we need them ( look awful atm)